Fix: Replace uses of `(void)` with `BSLA_` macros
This patch replaces uses of `(void)` with `BSLA_` macros, more accurately
marking variables as unused or potentially unused.  All cases where there were
casts to void were there to silence compiler warnings already, marking a
variable that is only used in an assertion as unused when assertions are turned
off.  The more modern way to do this is using the `BSLA_MAYBE_UNUSED`
annotation from BDE, which attaches to a variable declaration and tells the
compiler not to warn if the variable is unused.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>