Fix: Write logs on default `switch` cases
There are some cases where it would be catastrophic for the broker to
abort due to an unexpected value being provided in a `switch` statement.
This value may come over the network, which we have no control over, or
from a file on disk, which may have been corrupted.  This patch
introduces default cases to warn when this happens, but it keeps the
current behavior otherwise in these cases.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>