refactor: Remove unneeded `const_cast`
`mqba::ClientSession` uses a helper function `extractClientIdentity` to
return a pointer to the client identity from either a client’s
negotiation response or a broker’s negotiation response.  The client
identity that is returned is never modified, but `extractClientIdentity`
still has a `const_cast` to make this a pointer-to-non-`const`.  This
patch removes the `const_cast` and propagates `const`ness through the
return type of `extractClientIdentity`.

Signed-off-by: Patrick M. Niedzielski <pniedzielski@bloomberg.net>