Fix: Update assert message in `SessionOptions`
Both the precondition in the documentation for
`SessionOptions::setStatsDumpInterval` and the precondition assert
itself check that `value.seconds()` is less than 60 minutes, but the
message in the precondition assert claims it has to be less than 10
minutes.  This patch brings this message into line with our documented
and checked preconditions.

Signed-off-by: Patrick M. Niedzielski <pniedzielski@bloomberg.net>