Fix: Make compression ratio computation explicitly use `double`s
The compression ratio calculations in `bmqimp::Queue` use information
from the stats layer, which comes in `Int64`s, but does its computations
in `double`s.  This patch makes this conversion explicit, because it is
a narrowing conversion (there are `Int64`s which cannot be faithfully
represented as `double`s, even though we will never reach that size
integer).  There are two major changes:

  1. Make the compression ratio precision factor constant a `double`.
  2. Add explicit `static_cast`s from `Int64` to `double`s, and make all
     local variables `double`s rather than `Int64`s.

The combination of these changes fix a handful of warnings from GCC of
the sort:

    conversion from ‘long long int’ to ‘double’ may change value [-Wconversion]

Signed-off-by: Patrick M. Niedzielski <pniedzielski@bloomberg.net>